+++ src/js/index.js
@@ -1,4 +1,4 @@
-import {initRegister} from './register.js';
+import {initRegister} from './register/register.js';
 import {initRacing} from './racing.js';
 
 export const todoApp = () => {
